day5/python-sql-gce/app/requirements.txt
flask
psycopg2-binary
gunicorn
python-dotenv==1.0.1